- Felles knapp og logikk for toggle av varsler på mobil/desktop - Fjerner overflødige komponenter og css
