+++ src/api/api.ts
@@ -2,7 +2,7 @@
 import React from 'react';
 import { fetchToJson } from './api-utils';
 import { Data as innloggingsstatusData } from '../store/reducers/innloggingsstatus-duck';
-import { Data as varselinnboksData } from '../store/reducers/varselinnboks-duck';
+import { VarslerData as varselinnboksData } from '../store/reducers/varselinnboks-duck';
 import { MenyNode as menypunkterData } from '../store/reducers/menu-duck';
 
 export enum Status {
