+++ src/komponenter/header/header-regular/meny/MobilMenylinje.less
@@ -64,6 +64,11 @@
         margin-right: 3px;
     }
 
+    .varselbjelle__knapp {
+        order: 3;
+        margin-right: 0.5rem;
+    }
+
     .mobilmeny__knapp {
         order: 3;
         &__frame {
