+++ src/komponenter/header/header-regular/meny/ekspanderende-menyer/varsler-mobil/Varselvisning.less
@@ -1,32 +1,6 @@
 @import '../../../../../../../node_modules/nav-frontend-core/less/_variabler';
 
-.desktopmeny .min-varsel-wrapper,
 .mobilmeny .varsler-display-mobil-tablet {
-    .varsler-display-desktop {
-        display: block;
-        position: absolute;
-        top: 47px;
-        right: -37px;
-        left: auto;
-        width: 20rem;
-        padding: 1.25rem 1rem;
-        z-index: 1002;
-        background-color: white;
-        box-shadow: 0 1px 4px 2px rgba(0, 0, 0, 0.1);
-
-        &:before {
-            content: '';
-            position: absolute;
-            top: 0;
-            right: 1.625rem;
-            border: 0.5rem solid white;
-            z-index: 1003;
-            transform-origin: 0 0;
-            transform: rotate(135deg);
-            box-shadow: -2px 2px 3px 0 rgba(100, 100, 100, 0.1);
-        }
-    }
-
     .panel {
         display: block;
         border-radius: 0.25rem;
@@ -167,12 +141,6 @@
     .varsler-display-mobil-tablet {
         .vis-alle-lenke {
             padding: 1.5rem 0.75rem 0 3.125rem;
-        }
-    }
-
-    .varsler-display-desktop {
-        .vis-alle-lenke {
-            padding-top: 0.75rem;
         }
     }
 }
