--- src/komponenter/header/header-regular/meny/varsel/Varselbjelle.less
@@ -1,143 +1,0 @@
-@import '../../../../../../node_modules/nav-frontend-core/less/_variabler';
-@import '../../../../../styling-variabler';
-
-.mobilmeny .varselbjelle {
-    position: relative;
-    width: auto;
-    display: flex;
-    align-items: center;
-
-    .varsler-toggle-container {
-        display: inline-block;
-        border: 1px solid transparent;
-        vertical-align: middle;
-
-        .varselknapp-mobil {
-            display: flex;
-            align-items: center;
-            padding: 0.3rem 0.5rem;
-            min-width: 7.4375rem;
-        }
-
-        .word-wrapper {
-            display: inline-block;
-            position: relative;
-            text-align: left;
-
-            p {
-                opacity: 0;
-                display: inline-block;
-                position: absolute;
-                white-space: nowrap;
-                left: 0;
-                top: 0;
-            }
-            p.er-usynlig {
-                opacity: 0;
-            }
-
-            p.er-synlig {
-                opacity: 1;
-                position: relative;
-            }
-        }
-
-        @keyframes dytt-ut {
-            0% {
-                transform: translateX(0);
-            }
-            60% {
-                opacity: 0;
-                transform: translateX(110%);
-            }
-            100% {
-                opacity: 0;
-                transform: translateX(100%);
-            }
-        }
-        @keyframes dytt-inn {
-            0% {
-                opacity: 0;
-                transform: translateX(-100%);
-            }
-            60% {
-                opacity: 1;
-                transform: translateX(10%);
-            }
-            100% {
-                opacity: 1;
-                transform: translateX(0);
-            }
-        }
-
-        .er-synlig,
-        .er-usynlig {
-            animation-timing-function: ease;
-            animation-delay: 0s;
-            animation-iteration-count: 0;
-            animation-direction: normal;
-            animation-fill-mode: none;
-        }
-
-        .er-synlig {
-            animation: dytt-inn 0.6s;
-        }
-
-        .er-usynlig {
-            animation: dytt-ut 0.6s;
-        }
-
-        .varsler-toggle {
-            background: url('../../../../../ikoner/varsler/alarm.svg') no-repeat
-                center;
-            background-size: 1.5rem 1.5rem;
-            height: 2rem;
-            width: 2rem;
-            border: 2px solid transparent;
-
-            &:hover {
-                background-image: url('../../../../../ikoner/varsler/alarm_hover.svg');
-            }
-
-            &:active {
-                background-image: url('../../../../../ikoner/varsler/alarm_active.svg');
-            }
-
-            &:focus {
-                outline: 0;
-                border-radius: 4px;
-                background-image: url('../../../../../ikoner/varsler/alarm_active.svg');
-            }
-        }
-
-        &.har-nye-varsler .varsler-toggle {
-            background-image: url('../../../../../ikoner/varsler/alarm-nytt-varsel.svg');
-
-            &:hover {
-                background-image: url('../../../../../ikoner/varsler/alarm-nytt-varsel_hover.svg');
-            }
-
-            &:active {
-                background-image: url('../../../../../ikoner/varsler/alarm-nytt-varsel_active.svg');
-            }
-        }
-    }
-
-    @media (min-width: @tabletLimitWidth) {
-        width: 31px;
-
-        .varsler-toggle-container {
-            display: inline-block;
-            height: 1.9375rem;
-            width: 1.9375rem;
-
-            .varsler-toggle {
-                width: 100%;
-                height: 100%;
-                border: none;
-                padding: 0;
-                margin: 0;
-            }
-        }
-    }
-}
